Problem: there's no way to extend JMESPath
Yet, in some cases this is highly beneficial. I've done it in another
projects where custom extensions were used.

Solution: add the ability to implement custom functions

Other JMESPath implementations also support this feature, such as:

1. [jmespath.py](https://github.com/jmespath/jmespath.py#custom-functions)
2. [jmespath-ts](https://github.com/nanoporetech/jmespath-ts#register-you-own-custom-functions)
3. [jmespath for Rust](https://docs.rs/jmespath/0.2.0/jmespath/#custom-functions)

I haven't checked for me, but hopefully this is a good enough list to
confirm that this is not out of ordinary.